requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-ai-contentsafety
aiohttp
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-cosmos
semantic-kernel
tenacity